Initial upload
This was my first longer project using HTML and CSS. There is a bit of javascript used for the navbar popper function. Currently there is a bug with it, the navbar will expand but won't collapse. As I begin to understand more about JS, hopefully I can fix this. Also, the navbar href will sometimes automatically scroll to the section upon page refresh.  I suspect this has something to do with the cobbled together JS that I used to make the navbar popper function. I will try to fix this at a later date.